// src/common/types/chart-stats.ts
export type ChartStats = {
  uuid: string
  name: string
  date: string
  sport: string
  total_distance: number
  total_timer_time: number
  total_elapsed_time: number
  total_ascent: number
  avg_heart_rate: number
  enhanced_speed: number
  avg_tempo: number
  avg_timer_time: number
  avg_speed: number
  max_heart_rate: number
  total_calories: number
  avg_calories: number
  cadence_coef: number
}