// src/common/types/table-stats.ts
export type TableStats = {
  total_distance: number
  total_timer_time: number
  total_elapsed_time: number
  total_ascent: number
  avg_heart_rate: number
  avg_power: number
  avg_tempo: number
  avg_distance: number
  avg_timer_time: number
  avg_elapsed_time: number
  avg_speed: number
  avg_cadence: number
  max_heart_rate: number
  total_calories: number
  avg_calories: number
  count: number
}